feat: add service for exchange specs
