+++ README.md
@@ -16,5 +16,6 @@ python script_backtest.py --config configs/config_sim.yaml
 python script_train.py --config configs/train.yaml --trainer-module mypackage.trainer:MyTrainer
 python script_compare_runs.py run1 run2 run3            # сохранит compare_runs.csv
 python script_compare_runs.py run1 metrics.json --stdout  # вывод в stdout
+python script_fetch_exchange_specs.py --market futures --symbols BTCUSDT,ETHUSDT --out data/exchange_specs.json
 ```
